+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+You can provide a path to controller.analyzeImage(path) in order to scan a local photo from the gallery!
+Check out the example app to see how you can use the image_picker plugin to retrieve a photo from
+the gallery. Please keep in mind that this feature is only supported on Android and iOS.
+
 ## 0.1.3
 * Fixed crash after asking permission. [#29](https://github.com/juliansteenbakker/mobile_scanner/issues/29)
 * Upgraded cameraX from 1.1.0-beta01 to 1.1.0-beta02